markdown: Split out markdown_parse module.
We can't move this to shared unless we either configure
the linter to let markdown_parse import from static/third
or do some kind of function injection for the library.

We probably also want to wait until mobile has made some
progress on their side before actually sharing this.

This will end up creating a very minor amount of git
churn, but I think it makes sense to split the files up
now, even if it's just for the webapp to have a mutation
layer and a functional layer.

This is a pure code move, and the only things that change
are things like import statements.

For the tests I remove a completely unnecessary mock.
